# api_spec.yaml
openapi: 3.0.3
info:
  title: Organization Contributions API
  description: |
    REST API for tracking monthly employee contributions across products (Academy, Intensive, NIAT).
    
    ## Authentication
    All endpoints (except token generation) require JWT authentication. Include the access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Roles
    - **CEO**: Can access organization-wide dashboards
    - **HOD**: Can upload files and access department dashboards
    - **Pod Lead**: Can access pod-level dashboards
    - **Employee**: Can access own contribution data
    - **Admin**: Full access
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server
...
paths:
...
  /token/refresh/:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get a new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh
              properties:
                refresh:
                  type: string
                  example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/token/
...